tests/requirements.txt

pytest-mock==3.6.1
pytest-cov==2.12.1